[ MAJOR UPDATE ] restructured kritis3m_http functionality and modified sae_ID parameter from hardcoded to usage-dependent.
